feat: fix component name and contract params name
