+++ package.json
@@ -34,7 +34,7 @@
     "@ethersproject/bignumber": "^5.6.0",
     "@ethersproject/contracts": "^5.6.0",
     "@fontsource/dm-sans": "^4.5.7",
-    "@foxlottery/core-sdk": "^1.0.3",
+    "@foxlottery/core-sdk": "^1.0.4",
     "@gnosis.pm/safe-apps-web3-react": "^1.3.0",
     "@headlessui/react": "^1.6.0",
     "@heroicons/react": "^1.0.6",
