+++ src/config/routing.ts
@@ -1,5 +1,6 @@
 // a list of tokens by chain
 import { ChainId, Token, WNATIVE } from '@foxlottery/core-sdk'
+import ChainTokenList from 'app/types/ChainTokenList'
 
 import * as ARBITRUM from './tokens/arbitrum'
 import * as AVALANCHE from './tokens/avalanche'
@@ -15,13 +16,8 @@ import * as MOONBEAM from './tokens/moonbeam'
 import * as MOONRIVER from './tokens/moonriver'
 import * as OKEX from './tokens/okex'
 import * as PALM from './tokens/palm'
-import * as ROPSTEN from './tokens/ropsten'
 import * as TELOS from './tokens/telos'
 import * as XDAI from './tokens/xdai'
-
-type ChainTokenList = {
-  readonly [chainId: number]: Token[]
-}
 
 // List of all mirror's assets addresses.
 // Last pulled from : https://whitelist.mirror.finance/eth/tokenlists.json
@@ -44,7 +40,7 @@ const MIRROR_ADDITIONAL_BASES: { [tokenAddress: string]: Token[] } = {
   '0xf72FCd9DCF0190923Fadd44811E240Ef4533fc86': [ETHEREUM.MIR, ETHEREUM.UST], // mVIXY
 }
 
-const WRAPPED_NATIVE_ONLY: ChainTokenList = {
+export const WRAPPED_NATIVE_ONLY: ChainTokenList = {
   [ChainId.ETHEREUM]: [WNATIVE[ChainId.ETHEREUM]],
   [ChainId.ROPSTEN]: [WNATIVE[ChainId.ROPSTEN]],
   [ChainId.RINKEBY]: [WNATIVE[ChainId.RINKEBY]],
@@ -304,391 +300,6 @@ export const CUSTOM_BASES: {
   [ChainId.MATIC]: {
     [MATIC.TEL.address]: [MATIC.AAVE],
   },
-}
-
-/**
- * Shows up in the currency select for swap and add liquidity
- */
-export const COMMON_BASES: ChainTokenList = {
-  [ChainId.ETHEREUM]: [
-    ETHEREUM.WBTC,
-
-    ETHEREUM.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    ETHEREUM.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    ETHEREUM.DAI,
-  ],
-  [ChainId.ROPSTEN]: [ROPSTEN.DAI, ROPSTEN.LINK, ROPSTEN.UNI, ROPSTEN.WETH],
-  [ChainId.MATIC]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.MATIC],
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.MIM,
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.ICE,
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.DAI,
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.FRAX,
-    // @ts-ignore TYPE NEEDS FIXING
-    MATIC.STG,
-  ],
-  [ChainId.FANTOM]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.FANTOM],
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.MIM,
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.ICE,
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.SPELL,
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.DAI,
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.FRAX,
-    // @ts-ignore TYPE NEEDS FIXING
-    FANTOM.STG,
-  ],
-  [ChainId.BSC]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.BSC],
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.BTCB,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.MIM,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.SPELL,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.ICE,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.DAI,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.USD,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.FRAX,
-    // @ts-ignore TYPE NEEDS FIXING
-    BSC.STG,
-  ],
-  [ChainId.ARBITRUM]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.ARBITRUM],
-    // @ts-ignore TYPE NEEDS FIXING
-    ARBITRUM.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    ARBITRUM.MIM,
-    // @ts-ignore TYPE NEEDS FIXING
-    ARBITRUM.SPELL,
-    // @ts-ignore TYPE NEEDS FIXING
-    ARBITRUM.ICE,
-    // @ts-ignore TYPE NEEDS FIXING
-    ARBITRUM.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    ARBITRUM.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    ARBITRUM.FRAX,
-    // @ts-ignore TYPE NEEDS FIXING
-    ARBITRUM.STG,
-  ],
-  [ChainId.XDAI]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.XDAI],
-    // @ts-ignore TYPE NEEDS FIXING
-    XDAI.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    XDAI.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    XDAI.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    XDAI.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    XDAI.GNO,
-  ],
-  [ChainId.AVALANCHE]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.AVALANCHE],
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.ICE,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.MIM,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.SPELL,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.WMEMO,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.DAI,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.FRAX,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.STG,
-  ],
-  [ChainId.HARMONY]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.HARMONY],
-    // @ts-ignore TYPE NEEDS FIXING
-    HARMONY.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    HARMONY.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    HARMONY.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    HARMONY.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    HARMONY.DAI,
-    // @ts-ignore TYPE NEEDS FIXING
-    HARMONY.FRAX,
-  ],
-  [ChainId.HECO]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.HECO],
-    // @ts-ignore TYPE NEEDS FIXING
-    HECO.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    HECO.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    HECO.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    HECO.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    HECO.DAI,
-  ],
-  [ChainId.OKEX]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.OKEX],
-    // @ts-ignore TYPE NEEDS FIXING
-    OKEX.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    OKEX.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    OKEX.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    OKEX.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    OKEX.DAI,
-  ],
-  [ChainId.CELO]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.CELO],
-    // @ts-ignore TYPE NEEDS FIXING
-    CELO.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    CELO.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    CELO.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    CELO.cUSD,
-    // @ts-ignore TYPE NEEDS FIXING
-    CELO.cEURO,
-  ],
-  [ChainId.MOONRIVER]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.MOONRIVER],
-    // @ts-ignore TYPE NEEDS FIXING
-    MOONRIVER.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    MOONRIVER.BTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    MOONRIVER.MIM,
-    // @ts-ignore TYPE NEEDS FIXING
-    MOONRIVER.FRAX,
-    // @ts-ignore TYPE NEEDS FIXING
-    MOONRIVER.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    MOONRIVER.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    MOONRIVER.ROME,
-  ],
-  [ChainId.PALM]: [...WRAPPED_NATIVE_ONLY[ChainId.PALM], PALM.WETH, PALM.DAI],
-  [ChainId.FUSE]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.FUSE],
-    // @ts-ignore TYPE NEEDS FIXING
-    FUSE.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    FUSE.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    FUSE.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    FUSE.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    FUSE.DAI,
-  ],
-  [ChainId.TELOS]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.TELOS],
-    // @ts-ignore TYPE NEEDS FIXING
-    TELOS.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    TELOS.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    TELOS.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    TELOS.USDT,
-  ],
-  [ChainId.MOONBEAM]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.MOONBEAM],
-    MOONBEAM.USDC,
-    MOONBEAM.USDT,
-    MOONBEAM.UST,
-    MOONBEAM.WETH,
-    MOONBEAM.WBTC,
-    MOONBEAM.FRAX,
-  ],
-}
-
-// used to construct the list of all pairs we consider by default in the frontend
-export const BASES_TO_TRACK_LIQUIDITY_FOR: ChainTokenList = {
-  ...WRAPPED_NATIVE_ONLY,
-  [ChainId.ETHEREUM]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.ETHEREUM],
-    ETHEREUM.DAI,
-    ETHEREUM.USDC,
-    ETHEREUM.USDT,
-    ETHEREUM.WBTC,
-    ETHEREUM.OHM_V1,
-    ETHEREUM.OHM_V2,
-    ETHEREUM.MIM,
-    ETHEREUM.FRAX,
-    ETHEREUM.STG,
-  ],
-  [ChainId.MATIC]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.MATIC],
-    MATIC.USDC,
-    MATIC.WBTC,
-    MATIC.MIM,
-    MATIC.DAI,
-    MATIC.WETH,
-    MATIC.USDT,
-    MATIC.FRAX,
-    MATIC.STG,
-  ],
-  [ChainId.FANTOM]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.FANTOM],
-    FANTOM.DAI,
-    FANTOM.USDC,
-    FANTOM.WBTC,
-    FANTOM.WETH,
-    FANTOM.MIM,
-    FANTOM.FRAX,
-    FANTOM.STG,
-  ],
-  [ChainId.BSC]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.BSC],
-    BSC.DAI,
-    BSC.USD,
-    BSC.USDC,
-    BSC.USDT,
-    BSC.BTCB,
-    BSC.WETH,
-    BSC.MIM,
-    BSC.FRAX,
-    BSC.STG,
-  ],
-  [ChainId.ARBITRUM]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.ARBITRUM],
-    ARBITRUM.WBTC,
-    ARBITRUM.USDC,
-    ARBITRUM.USDT,
-    ARBITRUM.MIM,
-    ARBITRUM.gOHM,
-    ARBITRUM.MAGIC,
-    ARBITRUM.FRAX,
-    ARBITRUM.STG,
-  ],
-  [ChainId.XDAI]: [...WRAPPED_NATIVE_ONLY[ChainId.XDAI], XDAI.USDC, XDAI.USDT, XDAI.WBTC, XDAI.WETH],
-  [ChainId.AVALANCHE]: [
-    // @ts-ignore TYPE NEEDS FIXING
-    ...WRAPPED_NATIVE_ONLY[ChainId.AVALANCHE],
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.DAI,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.USDT,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.WBTC,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.WETH,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.USDC,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.MIM,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.TIME,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.FRAX,
-    // @ts-ignore TYPE NEEDS FIXING
-    AVALANCHE.STG,
-  ],
-  [ChainId.HARMONY]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.HARMONY],
-    HARMONY.DAI,
-    HARMONY.USDC,
-    HARMONY.USDT,
-    HARMONY.WBTC,
-    HARMONY.WETH,
-    HARMONY.FRAX,
-  ],
-  [ChainId.HECO]: [...WRAPPED_NATIVE_ONLY[ChainId.HECO], HECO.DAI, HECO.USDC, HECO.USDT, HECO.WBTC, HECO.WETH],
-  [ChainId.OKEX]: [...WRAPPED_NATIVE_ONLY[ChainId.OKEX], OKEX.DAI, OKEX.USDC, OKEX.USDT, OKEX.WBTC, OKEX.WETH],
-  [ChainId.CELO]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.CELO],
-    CELO.cETH,
-    CELO.mCUSD,
-    CELO.mCELO,
-    CELO.mcEURO,
-    CELO.cUSD,
-    CELO.cEURO,
-    CELO.cBTC,
-  ],
-  [ChainId.MOONRIVER]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.MOONRIVER],
-    MOONRIVER.USDC,
-    MOONRIVER.USDT,
-    MOONRIVER.WETH,
-    MOONRIVER.FRAX,
-    MOONRIVER.MIM,
-    MOONRIVER.BTC,
-    MOONRIVER.ROME,
-  ],
-  [ChainId.PALM]: [...WRAPPED_NATIVE_ONLY[ChainId.PALM], PALM.WETH, PALM.DAI],
-  [ChainId.FUSE]: [...WRAPPED_NATIVE_ONLY[ChainId.FUSE], FUSE.USDC, FUSE.USDT, FUSE.WBTC, FUSE.WETH, FUSE.DAI],
-  [ChainId.TELOS]: [...WRAPPED_NATIVE_ONLY[ChainId.TELOS], TELOS.USDC, TELOS.USDT, TELOS.WETH, TELOS.WBTC],
-  [ChainId.MOONBEAM]: [
-    ...WRAPPED_NATIVE_ONLY[ChainId.MOONBEAM],
-    MOONBEAM.USDC,
-    MOONBEAM.USDT,
-    MOONBEAM.UST,
-    MOONBEAM.WETH,
-    MOONBEAM.WBTC,
-    MOONBEAM.FRAX,
-  ],
 }
 
 export const PINNED_PAIRS: { readonly [chainId: number]: [Token, Token][] } = {
