+++ src/modals/CryptoCurrencyModal/index.tsx
@@ -1,7 +1,7 @@
 import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
+import CurrencyList from 'app/components/CurrencyList'
 import HeadlessUiModal from 'app/components/Modal/HeadlessUIModal'
-import CommonBases from 'app/modals/CryptoCurrencyModal/CommonBases'
 import { useCryptoCurrencyModalToggle, useIsModalOpen } from 'app/state/application/hooks'
 import { ApplicationModal } from 'app/state/application/reducer'
 import React, { FC } from 'react'
@@ -16,7 +16,7 @@ const CryptoCurrencyModal: FC = () => {
       <div className="flex flex-col gap-4">
         <HeadlessUiModal.Header header={i18n._(t`Select a Crypto Currency`)} onClose={toggleCryptoCurrencyModal} />
         <div className="grid grid-flow-row-dense grid-cols-1 gap-4 overflow-y-auto">
-          <CommonBases />
+          <CurrencyList />
         </div>
       </div>
     </HeadlessUiModal.Controlled>
