+++ src/types/ChainTokenList.ts
@@ -1,0 +1,7 @@
+import { Currency } from '@foxlottery/core-sdk'
+
+type ChainTokenList = {
+  readonly [chainId: number]: Currency[]
+}
+
+export default ChainTokenList
